Function obbdimal::bdd::bdd_ds::para_ite

fn para_ite(
    f: Arc<NodeType>, 
    g: Arc<NodeType>, 
    h: Arc<NodeType>, 
    unique_tables: Arc<Vec<Arc<Mutex<FnvHashMap<UniqueKey, Arc<NodeType>>>>>>, 
    cnf: &Cnf
) -> Arc<NodeType>

Calculates the ITE function. Basically works the same way as ite does, it just branches on every recursion step by using reyon::join.