Struct obbdimal::bdd::bdd_ds::Bdd

pub struct Bdd {
    unique_table: FnvHashMap<UniqueKey, Arc<NodeType>>,
    computed_table: FnvHashMap<ComputedKey, Arc<NodeType>>,
    cnf: Cnf,
    pub bdd: Arc<NodeType>,
}

Represents a wrapper struct for a BDD, allowing us to query methods on it.

Fields
unique_table: FnvHashMap<UniqueKey, Arc<NodeType>>computed_table: FnvHashMap<ComputedKey, Arc<NodeType>>cnf: Cnfbdd: Arc<NodeType>
Implementations

impl Bdd

pub(crate) fn from_format(
data: &str,
format: InputFormat,
settings: ParserSettings,
static_ordering: StaticOrdering
) -> Result<Self, DataFormatError>

Creates a new instance of a Bdd in a sequential fashion out of a given input format. Currently there is only InputFormat::CNF supported, which represents Dimacs CNF.

pub(crate) fn from_format_para(
data: &str,
format: InputFormat,
settings: ParserSettings,
static_ordering: StaticOrdering
) -> Result<Self, DataFormatError>

Creates a new instance of a Bdd in a parallelized fashion out of a given input format. Currently there is only InputFormat::CNF supported, which represents Dimacs CNF.

fn from_cnf_para(
symbols: Symbol,
cnf: &Cnf,
unique_tables: Arc<Vec<Arc<Mutex<FnvHashMap<UniqueKey, Arc<NodeType>>>>>>
) -> Self

Creates a new instance of a BDD manager from a given CNF.

fn from_cnf_para_rec(
symbols: Symbol,
cnf: &Cnf,
unique_tables: Arc<Vec<Arc<Mutex<FnvHashMap<UniqueKey, Arc<NodeType>>>>>>
) -> Arc<NodeType>

Helper method for from_cnf_para.

fn from_cnf(symbols: Symbol, cnf: Cnf) -> Self

Creates a new instance of a BDD manager from a given CNF.

fn from_cnf_rec(&mut self, cnf: Symbol) -> Arc<NodeType>

Helper method for from_cnf.

fn add_node_to_unique(
&mut self,
var: i64,
low: Arc<NodeType>,
high: Arc<NodeType>
) -> Arc<NodeType>

Adds a NodeType to the unique_table, if it is not already there.

pub(crate) fn nodecount(&self) -> u64

Returns the number of nodes in the Bdd.

fn nodecount_rec(subtree: Arc<NodeType>) -> u64

This is the helper function for nodecount. This function just iterates through all nodes and counts them.

pub(crate) fn satcount(&self) -> u64

Returns the number of variable assignments that evaluate the represented BDD to true.

pub(crate) fn satisfiable(&self) -> bool

Returns true if there is a variable assignment which evaluates the represented formula to true.

fn restrict(
&mut self,
node: Arc<NodeType>,
v: i64,
order: &Vec<i32>,
val: bool
) -> Arc<NodeType>

Applies either true or false val to the children of a given node. It's basically the Shannon decomposition.

fn ite(
&mut self,
f: Arc<NodeType>,
g: Arc<NodeType>,
h: Arc<NodeType>
) -> Arc<NodeType>

If-then-else, if f ite returns g, else h.

pub(crate) fn and(
&mut self,
lhs: Arc<NodeType>,
rhs: Arc<NodeType>
) -> Arc<NodeType>

Calculates the Boolean AND with the given left hand side lhs and the given right hand side rhs.

pub(crate) fn or(
&mut self,
lhs: Arc<NodeType>,
rhs: Arc<NodeType>
) -> Arc<NodeType>

Calculates the Boolean OR with the given left hand side lhs and the given right hand side rhs.

pub(crate) fn not(&mut self, val: Arc<NodeType>) -> Arc<NodeType>

Calculates the Boolean NOT with the given value val.

pub(crate) fn serialize(&self) -> String

Serializes self to a String representing the BDD. The serialization of the BDD obeys the following rules:

The first line of the string is the variable ordering of the BDD.

Every following line represents a node, where the first number is the internal ID of the node the second number is the top_var of the current node, the thrid number is the id of the node connected to the low edge the fourth number is the id of the node connected to the high edge of the current node.

Internal ID 0 and 1 are representations of the terminal ZERO and ONE node. The last two lines have to be appended as the sink nodes for the recursive deserialization to work.

fn serialize_rec(subtree: Arc<NodeType>) -> String

pub(crate) fn deserialize(input: String) -> Bdd

Deserializes the given string (which was previously serialized by serialize) into a Bdd. TODO: Error handling for wrong input formats. TODO: Enhance runtime by removing linear iteration of input data.

fn deserialize_rec(
complete_input: String,
current: String,
hash_map: FnvHashMap<UniqueKey, Arc<NodeType>>
) -> (Arc<NodeType>, FnvHashMap<UniqueKey, Arc<NodeType>>)

Helper function from deserialize.