Module obbdimal::bdd::bdd_graph

Structs

Node
Representation of a Binary Decision Diagram node, containing the top variable top_var and the children hi and lo.

Enums

NodeType
Representation of what types a Node in a BDD can be. Zero is the terminal 0 node. One is the terminal 1 node. Complex(Node) represents a Node.

Statics

NODE_ID