Trait obbdimal::bdd::manager::Manager

pub trait Manager {
    fn get_bdd(&self) -> &Option<Bdd>;

    fn get_sat_count(&self) -> &Option<u64>;

    fn get_node_count(&self) -> &Option<u64>;

    fn get_bdd_mut(&mut self) -> &mut Option<Bdd>;

    fn get_sat_count_mut(&mut self) -> &mut Option<u64>;

    fn get_node_count_mut(&mut self) -> &mut Option<u64>;

    fn new() -> Self;

    fn from_format(
        cnf: &str, 
        format: InputFormat, 
        settings: ParserSettings, 
        static_ordering: StaticOrdering
    ) -> Result<Self, DataFormatError>
    where
        Self: Sized;

    fn add_bdd(&mut self, bdd: Bdd) -> &mut Self { ... }

    fn deserialize_bdd(&mut self, input: &str) -> &mut Self { ... }

    fn serialize_bdd(&self) -> Result<String, NoBddError> { ... }

    fn node_count(&mut self) -> Result<u64, NoBddError> { ... }

    fn satisfiable(&self) -> Result<bool, NoBddError> { ... }

    fn sat_count(&mut self) -> Result<u64, NoBddError> { ... }
}

Holds all the methods a Manager should have. To give default behaviour to the calculation methods every Manager has to implement those getters.

Required methods

fn get_bdd(&self) -> &Option<Bdd>

fn get_sat_count(&self) -> &Option<u64>

fn get_node_count(&self) -> &Option<u64>

fn get_bdd_mut(&mut self) -> &mut Option<Bdd>

fn get_sat_count_mut(&mut self) -> &mut Option<u64>

fn get_node_count_mut(&mut self) -> &mut Option<u64>

fn new() -> Self

Creates an empty 'Manager' struct.

fn from_format(
cnf: &str,
format: InputFormat,
settings: ParserSettings,
static_ordering: StaticOrdering
) -> Result<Self, DataFormatError> where
Self: Sized,

Creates a new BDD from a given input. format is the given format of the input. settings describe how the parser should interpret the input. Returns a BddManager or a DataFormatError.

Loading content...
Provided methods

fn add_bdd(&mut self, bdd: Bdd) -> &mut Self

Adds a given Bdd to Manager and resets the memoization of the sat_count and node_count.

fn deserialize_bdd(&mut self, input: &str) -> &mut Self

Deserializes a previously serialized Bdd onto the Manager.

fn serialize_bdd(&self) -> Result<String, NoBddError>

Serializes the current hold Bdd to a String.

fn node_count(&mut self) -> Result<u64, NoBddError>

Counts the nodes of the currently hold Bdd and returns the result Result<u64, NoBddError>.

fn satisfiable(&self) -> Result<bool, NoBddError>

Returns Ok(true) is the given Bdd represents a function which is satisfiable.

fn sat_count(&mut self) -> Result<u64, NoBddError>

Returns the number of inputs that satisfy the function the current Bdd is representing.

Loading content...