Module obbdimal::input::boolean_function

Structs

BooleanFunction
Represents a Boolean function.

Enums

Operator
Represents all the operations currently supported by the BooleanFunction struct.

Symbol
A Symbol represents either a BooleanFunction, a terminal symbol containing the index of a variable Posterminal(u32) or a terminal symbol containing the index of a negated variable Negterminal(u32)