Module obbdimal::input::static_ordering

Enums

StaticOrdering
Currently supported heurisitcs for static variable ordering.

Functions

apply_force

apply_heuristic
Applies a given static variable ordering heuristic to a given Cnf and returns a new Cnf with the newly calculated order. Currently the following heuristics are supported: -NONE -FORCE